+++ src/bayonetta/acmd/other.rs
@@ -13,3 +13,7 @@ use smash::app::*;
 use smash::phx::Vector3f;
 use crate::util::*;
 use super::*;
+
+pub fn install() {
+    
+}
